Snowflake/SourceCode/check_env.py
import os
from dotenv import load_dotenv

load_dotenv()

keys = [
    "SNOWFLAKE_ACCOUNT",
    "SNOWFLAKE_USER",
    "SNOWFLAKE_PASSWORD",
    "SNOWFLAKE_ROLE",
    "SNOWFLAKE_WAREHOUSE",
    "SNOWFLAKE_DATABASE",
    "SNOWFLAKE_SCHEMA"
]

for key in keys:
    print(f"{key}: {os.getenv(key)}")
