Snowflake/SourceCode/requirements.txt
snowflake-snowpark-python
python-dotenv
